Updated project namespace to match spelling of Dompdf project.
