+++ Module.php
@@ -17,4 +17,4 @@
  * @license	http://www.opensource.org/licenses/mit-license.php MIT License
  */
 
-require_once __DIR__ . '/src/DOMPDFModule/Module.php';
+require_once __DIR__ . '/src/DompdfModule/Module.php';
