+++ src/DompdfModule/Mvc/Service/ViewPdfStrategyFactory.php
@@ -17,19 +17,28 @@
  * @license	http://www.opensource.org/licenses/mit-license.php MIT License
  */
 
-use DompdfModuleTest\Framework\TestCase;
+namespace DompdfModule\Mvc\Service;
 
-ini_set('display_errors', 'On');
-error_reporting(E_ALL | E_STRICT);
+use Zend\ServiceManager\Factory\FactoryInterface;
+use Interop\Container\ContainerInterface;
+use DompdfModule\View\Strategy\PdfStrategy;
 
-if (is_readable(__DIR__ . '/TestConfiguration.php')) {
-    $configuration = include_once __DIR__ . '/TestConfiguration.php';
-} else {
-    $configuration = include_once __DIR__ . '/TestConfiguration.php.dist';
+class ViewPdfStrategyFactory implements FactoryInterface
+{
+    /**
+     * Create and return the PDF view strategy
+     *
+     * Retrieves the ViewPdfRenderer service from the service locator, and
+     * injects it into the constructor for the PDF strategy.
+     *
+     * @SuppressWarnings("unused")
+     * @param ContainerInterface $container
+     * @param string $requestedName
+     * @param array|null $options
+     * @return PdfStrategy
+     */
+    public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
+    {
+        return new PdfStrategy($container->get('ViewPdfRenderer'));
+    }
 }
-
-require_once __DIR__ . '/../vendor/autoload.php';
-
-$application = \Zend\Mvc\Application::init($configuration);
-$serviceManager = $application->getServiceManager();
-TestCase::setServiceManager($serviceManager);
