+++ tests/DompdfModuleTest/ModuleTest.php
@@ -17,19 +17,39 @@
  * @license	http://www.opensource.org/licenses/mit-license.php MIT License
  */
 
-use DompdfModuleTest\Framework\TestCase;
+namespace DompdfModuleTest;
 
-ini_set('display_errors', 'On');
-error_reporting(E_ALL | E_STRICT);
+use DompdfModule\Module;
 
-if (is_readable(__DIR__ . '/TestConfiguration.php')) {
-    $configuration = include_once __DIR__ . '/TestConfiguration.php';
-} else {
-    $configuration = include_once __DIR__ . '/TestConfiguration.php.dist';
+class ModuleTest extends \PHPUnit_Framework_TestCase
+{
+    /**
+     * @var Module
+     */
+    private $module;
+
+    public function testHasConfig()
+    {
+        $config = $this->module->getConfig();
+
+        // Test the obvious required keys.
+        $this->assertArrayHasKey('dompdf_module', $config, 'dompdf_module');
+        $this->assertArrayHasKey('service_manager', $config, 'service_manager');
+    }
+
+    public function testHasAutoloaderConfig()
+    {
+        $config = $this->module->getAutoloaderConfig();
+        $this->assertInternalType('array', $config, 'config is array');
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    protected function setUp()
+    {
+        parent::setUp();
+
+        $this->module = new Module();
+    }
 }
-
-require_once __DIR__ . '/../vendor/autoload.php';
-
-$application = \Zend\Mvc\Application::init($configuration);
-$serviceManager = $application->getServiceManager();
-TestCase::setServiceManager($serviceManager);
